requirements.txt
pubmed-parser
sentence-transformers
vertexai
google-cloud-aiplatform
chromadb
pydantic
uvicorn